fix(core): Module detection did not work in Node 12+
The behavior of `require.resolve` has apparently changed in Node 12+,
rendering it unable to find `package.json` files. It is possible this is
a bug in Node 12+, however I'm not certain.

Replaced the behavior with a simpler lookup mechanism that appears to
work, and should be expected to continue working at least for those
packages we care about (we control the file layout of those...)

A similar issue was affecting applet loading, and it was fixed in a
similar way.